backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib[bcrypt]
python-jose
python-multipart
PyJWT
cryptography
bcrypt==3.2.2

